More touchups.

+++ Integration Programs/ScannerAudioIntegration/ScannerAudioIntegration.ino
@@ -74,7 +74,8 @@ char uid[32];                                                                   
 /// </summary>
 void IRAM_ATTR ISR_volumeUp()
 {
-  if (volume < 30 && !upPressed) {
+  if (volume < 30 && !upPressed)
+  {
     volume = volume + 3;  // Maximum volume is 30, check underneath.
     volume = volume > 30 ? 30 : volume; // Ensures volume is within the dedicated range. Redundancy.
     upPressed = true;
@@ -87,11 +88,12 @@ void IRAM_ATTR ISR_volumeUp()
 /// </summary>
 void IRAM_ATTR ISR_volumeDown()
 {
-  if (volume > 0 && !downPressed) {
-      volume = volume - 3;  // Ninimum volume is 0, check underneath.
-      volume = volume < 0 ? 0 : volume; // Ensures volume is within the dedicated range. Redundancy.
-      downPressed = true;
-    }
+  if (volume > 0 && !downPressed)
+  {
+    volume = volume - 3;  // Ninimum volume is 0, check underneath.
+    volume = volume < 0 ? 0 : volume; // Ensures volume is within the dedicated range. Redundancy.
+    downPressed = true;
+  }
 }
 
 void setup()
@@ -139,9 +141,12 @@ void loop()
 {
   checkVolume();
   // Check if the player has finished playing the current file
-  if (player.available()) {
-    int type = player.readType();
-    if (type == DFPlayerPlayFinished) {
+  if (player.available())
+  {
+    uint8_t type = player.readType();
+
+    if (type == DFPlayerPlayFinished)
+    {
       Serial.println("Track finished!");
       isPlaying = false;
     }
